+++ standalone/include/psen_scan_v2_standalone/laserscan.h
@@ -32,9 +32,10 @@ namespace psen_scan_v2_standalone
  *
  * - Distance measures in mm.
  * - Normalized intensities of the signals.
- * - Resolution of the scan in radians.
- * - Min and Max angle in radians.
+ * - Resolution of the scan.
+ * - Min and Max angle.
  * - Counter of scan round.
+ * - ID of the currently active zoneset.
  * - Time of the first scan ray.
  *
  * The measures use the target frame defined as \<tf_prefix\>.
@@ -51,6 +52,7 @@ public:
             const util::TenthOfDegree& min_scan_angle,
             const util::TenthOfDegree& max_scan_angle,
             const uint32_t scan_counter,
+            const uint8_t active_zoneset,
             const int64_t timestamp);
 
 public:
@@ -58,6 +60,7 @@ public:
   const util::TenthOfDegree& getMinScanAngle() const;
   const util::TenthOfDegree& getMaxScanAngle() const;
   uint32_t getScanCounter() const;
+  uint8_t getActiveZoneset() const;
   int64_t getTimestamp() const;
 
   const MeasurementData& getMeasurements() const;
@@ -72,14 +75,16 @@ private:
   MeasurementData measurements_;
   //! Stores the received normalized signal intensities.
   IntensityData intensities_;
-  //! Distance of angle between the measurements (in radian).
+  //! Distance of angle between the measurements.
   const util::TenthOfDegree resolution_;
-  //! Lowest angle the scanner is scanning (in radian).
+  //! Lowest angle the scanner is scanning.
   const util::TenthOfDegree min_scan_angle_;
-  //! Highest angle the scanner is scanning (in radian).
+  //! Highest angle the scanner is scanning.
   const util::TenthOfDegree max_scan_angle_;
   //! Number of the scan round this data belongs to.
   const uint32_t scan_counter_;
+  //! The currently active zoneset of the scanner.
+  const uint8_t active_zoneset_;
   //! Time of the first ray in this scan round (or fragment if fragmented_scans is enabled).
   const int64_t timestamp_;
 };