+++ standalone/src/data_conversion_layer/monitoring_frame_deserialization.cpp
@@ -112,6 +112,18 @@ monitoring_frame::Message deserialize(const data_conversion_layer::RawData& data
         end_of_frame = true;
         break;
 
+      case AdditionalFieldHeaderID::zone_set:
+        if (additional_header.length() != NUMBER_OF_BYTES_ZONE_SET)
+        {
+          throw ZoneSetUnexpectedSize(fmt::format("Length of zone set field is {}, but should be {}.",
+                                                  additional_header.length(),
+                                                  NUMBER_OF_BYTES_ZONE_SET));
+        }
+        uint8_t zone_set_read_buffer;
+        raw_processing::read<uint8_t>(ss, zone_set_read_buffer);
+        msg.active_zoneset_ = zone_set_read_buffer;
+        break;
+
       case AdditionalFieldHeaderID::diagnostics:
         msg.diagnostic_messages_ = diagnostic::deserializeMessages(ss);
         msg.diagnostic_data_enabled_ = true;