+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Forthcoming
 * Always build hardware tests
 * Internal refactorings
 * Importing the config xml-file and publishing the zonesets
+* Add active zoneset to LaserScan
+* Publish active zoneset id to ~/active_zoneset
 
 0.3.2 (2021-09-16)
 ------------------